extract utilities_gui.py command line UI from kohya_gui/utilities.py
