+++ kohya_gui/utilities.py
@@ -5,7 +5,7 @@
 
 import gradio as gr
 import os
-import argparse
+
 from .basic_caption_gui import gradio_basic_caption_gui_tab
 from .convert_model_gui import gradio_convert_model_tab
 from .blip_caption_gui import gradio_blip_caption_gui_tab
@@ -38,61 +38,4 @@ def utilities_tab(
         reg_data_dir_input,
         output_dir_input,
         logging_dir_input,
-    )
-
-
-def UI(**kwargs):
-    css = ''
-
-    if os.path.exists('./style.css'):
-        with open(os.path.join('./style.css'), 'r', encoding='utf8') as file:
-            print('Load CSS...')
-            css += file.read() + '\n'
-
-    interface = gr.Blocks(css=css)
-
-    with interface:
-        utilities_tab()
-
-    # Show the interface
-    launch_kwargs = {}
-    if not kwargs.get('username', None) == '':
-        launch_kwargs['auth'] = (
-            kwargs.get('username', None),
-            kwargs.get('password', None),
-        )
-    if kwargs.get('server_port', 0) > 0:
-        launch_kwargs['server_port'] = kwargs.get('server_port', 0)
-    if kwargs.get('inbrowser', False):
-        launch_kwargs['inbrowser'] = kwargs.get('inbrowser', False)
-    print(launch_kwargs)
-    interface.launch(**launch_kwargs)
-
-
-if __name__ == '__main__':
-    # torch.cuda.set_per_process_memory_fraction(0.48)
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        '--username', type=str, default='', help='Username for authentication'
-    )
-    parser.add_argument(
-        '--password', type=str, default='', help='Password for authentication'
-    )
-    parser.add_argument(
-        '--server_port',
-        type=int,
-        default=0,
-        help='Port to run the server listener on',
-    )
-    parser.add_argument(
-        '--inbrowser', action='store_true', help='Open in browser'
-    )
-
-    args = parser.parse_args()
-
-    UI(
-        username=args.username,
-        password=args.password,
-        inbrowser=args.inbrowser,
-        server_port=args.server_port,
     )
